hopeSystem/requirements.txt
flask
flask-cors
mysql-connector-python
python-dotenv
pip install paddlepaddle paddleocr

pip install easyocr -i https://pypi.tuna.tsinghua.edu.cn/simple --user